Implementation of a PokeHandler class to work on collections of Pokemon.
Also makes static, single-target functions available publicly so it can
also be used as a static function collection to work on Pokemon.

This version implements rename with patterns, bulk rename with patterns
and the method `getLocalPokeName`, which is already implemented in
BlossomsPoGoManager.java as `getPokemonName` (TODO: remove it there and replace it's calls
to the new static method).