Modify "Evolvable" column to display both the required candy count _without_ and candy count _with_ "Transfer After Evolve" enabled, separated by a "/". The use of a separator character requires the column type to be changed from NULLABLE_INT to STRING.
However, this poses a problem when sorting as this would position 10's and 20's, etc. together with 1's and 2's and so forth. To prevent this the first value in front of the "/" character needs to be treated as an integer type during sort.

To allow for this, a new column type "NUMBER_STRING" is defined and linked to a dedicated comparator. This comparator then looks for the "/" character and if found, splits the string into two parts of which the first part is than converted back to integer during the sort comparison.

The split variables receiving the separated values are pre-initialized with "-1" for the parts preceding a possibly present "/", so that for pokémon which cannot be evolved, the NO_VALUE_SIGN is sorted separately from those with a zero count.